chore: fix final build and lint blockers
